# Backend/app/Http/Controllers/Api/CheckoutController.php
<?php

namespace App\Http\Controllers\Api;

use App\Services\SocketService;
use App\Http\Controllers\Controller;
use App\Mail\OrderReceipt;
use App\Models\CartItem;
use App\Models\Notification;
use App\Models\Order;
use App\Models\OrderItem;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Mail;
use Illuminate\Support\Facades\Validator;
use OpenApi\Attributes as OA;

class CheckoutController extends Controller
{
    #[OA\Post(
        path: '/api/checkout',
        summary: 'Place an order from the user cart',
        security: [['sanctum' => []]],
        tags: ['Checkout'],
        requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'payment_method', type: 'string', enum: ['aba']),
                new OA\Property(property: 'shipping_name', type: 'string'),
                new OA\Property(property: 'shipping_phone', type: 'string'),
                new OA\Property(property: 'shipping_address', type: 'string'),
                new OA\Property(property: 'shipping_city', type: 'string'),
                new OA\Property(property: 'shipping_zip', type: 'string'),
            ]
        )),
        responses: [
            new OA\Response(response: 201, description: 'Order placed successfully'),
            new OA\Response(response: 422, description: 'Validation error'),
        ]
    )]
    public function checkout(Request $request)
    {
        $cartItems = CartItem::with('product')
            ->where('user_id', $request->user()->id)
            ->get();

        if ($cartItems->isEmpty()) {
            return response()->json(['message' => 'Your cart is empty.'], 422);
        }

        $validator = Validator::make($request->all(), [
            'payment_method' => 'required|string|in:aba',
            'shipping_name'    => 'nullable|string|max:255',
            'shipping_phone'   => 'nullable|string|max:20',
            'shipping_address' => 'nullable|string|max:500',
            'shipping_city'    => 'nullable|string|max:255',
            'shipping_zip'     => 'nullable|string|max:20',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'message' => 'Please select a payment method.',
                'errors'  => $validator->errors(),
            ], 422);
        }

        $errors = [];
        foreach ($cartItems as $item) {
            if (! $item->product || ! $item->product->is_active) {
                $name = $item->product?->name ?? 'Unknown product';
                $errors[] = "'{$name}' is no longer available.";
            } elseif ($item->quantity > $item->product->stock) {
                $name = $item->product->name ?? 'Unknown product';
                $errors[] = "'{$name}' only has {$item->product->stock} units available.";
            }
        }

        if (! empty($errors)) {
            return response()->json(['message' => 'Some items cannot be ordered.', 'errors' => $errors], 422);
        }

        $order = DB::transaction(function () use ($cartItems, $request) {
            $subtotal = $cartItems->sum(fn ($item) => $item->product->price * $item->quantity);
            $tax = round($subtotal * 0.1, 2);
            $total = round($subtotal + $tax, 2);

            $order = Order::create([
                'user_id'          => $request->user()->id,
                'total'            => $total,
                'status'           => 'pending',
                'payment_method'   => $request->payment_method,
                'payment_status'   => 'unpaid',
                'shipping_name'    => $request->shipping_name,
                'shipping_phone'   => $request->shipping_phone,
                'shipping_address' => $request->shipping_address,
                'shipping_city'    => $request->shipping_city,
                'shipping_zip'     => $request->shipping_zip,
            ]);

            foreach ($cartItems as $item) {
                OrderItem::create([
                    'order_id'   => $order->id,
                    'product_id' => $item->product_id,
                    'quantity'   => $item->quantity,
                    'price'      => $item->product->price,
                    'status'     => 'pending',
                ]);

                $item->product->decrement('stock', $item->quantity);
            }

            CartItem::where('user_id', $request->user()->id)->delete();

            return $order;
        });

        $order->load(['user', 'items.product']);

        try {
            Mail::to($order->user->email)->send(new OrderReceipt($order));
        } catch (\Throwable $e) {
            // Log email failure but don't block the order
        }

        try {
            $notif = Notification::create([
                'title'   => 'Order #' . $order->id . ' Placed',
                'message' => 'Your order of $' . number_format($order->total, 2) . ' has been placed successfully.',
                'type'    => 'news',
                'link'    => '/orders/' . $order->id . '/receipt',
            ]);
            $notif->reads()->attach($request->user()->id, ['read_at' => null]);

            $adminIds = User::whereHas('role', fn ($q) => $q->where('name', 'admin'))
                ->pluck('id');
            $notif->reads()->attach($adminIds->toArray(), ['read_at' => null]);

            SocketService::notification([
                'id'         => $notif->id,
                'title'      => $notif->title,
                'message'    => $notif->message,
                'type'       => $notif->type,
                'link'       => $notif->link,
                'created_at' => $notif->created_at->toIso8601String(),
                'user_id'    => $request->user()->id,
            ]);

            SocketService::push('admin-notification', [
                'id'         => $notif->id,
                'title'      => 'New Order #' . $order->id,
                'message'    => 'Customer ' . $order->user->name . ' placed an order of $' . number_format($order->total, 2),
                'type'       => 'order',
                'link'       => '/admin/orders/' . $order->id . '/edit',
                'created_at' => $notif->created_at->toIso8601String(),
                'order_id'   => $order->id,
                'customer'   => $order->user->name,
                'total'      => $order->total,
            ]);

            SocketService::cartUpdate();
        } catch (\Throwable $e) {
            // Don't block order if notification fails
        }

        return response()->json([
            'message' => 'Order placed successfully.',
            'order' => $order,
        ], 201);
    }

    #[OA\Post(
        path: '/api/orders/{order}/payment-proof',
        summary: 'Upload payment proof for an order',
        security: [['sanctum' => []]],
        tags: ['Checkout'],
        parameters: [
            new OA\Parameter(name: 'order', in: 'path', required: true, schema: new OA\Schema(type: 'integer')),
        ],
        requestBody: new OA\RequestBody(required: true, content: new OA\MediaType(
            mediaType: 'multipart/form-data',
            schema: new OA\Schema(
                properties: [
                    new OA\Property(property: 'payment_proof', type: 'string', format: 'binary'),
                ]
            )
        )),
        responses: [
            new OA\Response(response: 200, description: 'Payment proof uploaded'),
            new OA\Response(response: 403, description: 'Forbidden'),
            new OA\Response(response: 422, description: 'Validation error'),
        ]
    )]
    public function uploadProof(Request $request, Order $order)
    {
        if ($order->user_id !== $request->user()->id) {
            return response()->json(['message' => 'Forbidden'], 403);
        }

        if ($order->payment_status !== 'unpaid') {
            return response()->json(['message' => 'Payment already submitted.'], 422);
        }

        $validator = Validator::make($request->all(), [
            'payment_proof' => 'required|image|mimes:jpeg,png,jpg|max:5120',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'message' => 'Please upload a valid payment screenshot (jpg/png, max 5MB).',
                'errors'  => $validator->errors(),
            ], 422);
        }

        $path = $request->file('payment_proof')->store('payments', 'public');

        $order->update([
            'payment_proof' => $path,
            'payment_status' => 'pending_verification',
        ]);

        $order->load(['user', 'items.product']);

        return response()->json([
            'message' => 'Payment proof uploaded. Awaiting verification.',
            'order'   => $order,
        ]);
    }
}