# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Shop System API",
        "description": "API documentation for the Online Shop System. Supports public, customer, and admin endpoints.",
        "contact": {
            "email": "support@onlineshop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/checkout": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Place an order from the user cart",
                "operationId": "fbc3557c7536e14478957c16f5199ef8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "aba"
                                        ]
                                    },
                                    "shipping_name": {
                                        "type": "string"
                                    },
                                    "shipping_phone": {
                                        "type": "string"
                                    },
                                    "shipping_address": {
                                        "type": "string"
                                    },
                                    "shipping_city": {
                                        "type": "string"
                                    },
                                    "shipping_zip": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}/payment-proof": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Upload payment proof for an order",
                "operationId": "bcd5512a53fbac2388a283b126451ac9",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "payment_proof": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment proof uploaded"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications for authenticated user",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "responses": {
                    "200": {
                        "description": "List of notifications with unread count"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{notification}/read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "operationId": "72e815431bf7f79bc33df9bcc2d4d957",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked as read"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "fc83170240b5026140e1ab9ffc095173",
                "responses": {
                    "200": {
                        "description": "All marked as read"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/toggle": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Toggle notification subscription",
                "operationId": "508e37463a8d1bb6a7222bf2858874d0",
                "responses": {
                    "200": {
                        "description": "Subscription toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders for authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Paginated list of orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token obtained from login/register. Format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}